# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Allied School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: class
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /parents/child-progress/{rollNumber}:
    get:
      operationId: getChildProgress
      tags: [students]
      summary: Get child progress by roll number (parent portal)
      parameters:
        - name: rollNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Child progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChildProgress"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Student:
      type: object
      properties:
        id:
          type: integer
        rollNumber:
          type: string
        name:
          type: string
        class:
          type: string
        section:
          type: ["string", "null"]
        fatherName:
          type: string
        motherName:
          type: ["string", "null"]
        phone:
          type: string
        address:
          type: ["string", "null"]
        admissionDate:
          type: string
        monthlyFee:
          type: integer
        isActive:
          type: boolean
        leftDate:
          type: ["string", "null"]
      required: [id, rollNumber, name, class, fatherName, phone, admissionDate, monthlyFee, isActive]
    StudentDetail:
      type: object
      properties:
        id:
          type: integer
        rollNumber:
          type: string
        name:
          type: string
        class:
          type: string
        section:
          type: ["string", "null"]
        fatherName:
          type: string
        motherName:
          type: ["string", "null"]
        phone:
          type: string
        address:
          type: ["string", "null"]
        admissionDate:
          type: string
        monthlyFee:
          type: integer
        isActive:
          type: boolean
        leftDate:
          type: ["string", "null"]
        feeHistory:
          type: array
          items:
            $ref: "#/components/schemas/FeeRecord"
        attendanceSummary:
          $ref: "#/components/schemas/AttendanceSummary"
      required: [id, rollNumber, name, class, fatherName, phone, admissionDate, monthlyFee, isActive, feeHistory, attendanceSummary]
    StudentInput:
      type: object
      properties:
        rollNumber:
          type: string
        name:
          type: string
        class:
          type: string
        section:
          type: string
        fatherName:
          type: string
        motherName:
          type: string
        phone:
          type: string
        address:
          type: string
        admissionDate:
          type: string
        monthlyFee:
          type: integer
      required: [rollNumber, name, class, fatherName, phone, admissionDate, monthlyFee]
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
        class:
          type: string
        section:
          type: string
        fatherName:
          type: string
        motherName:
          type: string
        phone:
          type: string
        address:
          type: string
        monthlyFee:
          type: integer
        isActive:
          type: boolean
        leftDate:
          type: string
    FeeRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        class:
          type: ["string", "null"]
        month:
          type: integer
        year:
          type: integer
        amount:
          type: integer
        paid:
          type: boolean
        paidDate:
          type: ["string", "null"]
      required: [id, studentId, month, year, amount, paid]
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        class:
          type: ["string", "null"]
        date:
          type: string
        present:
          type: boolean
      required: [id, studentId, date, present]
...
    AttendanceSummary:
      type: object
      properties:
        totalDays:
          type: integer
        presentDays:
          type: integer
        absentDays:
          type: integer
        percentage:
          type: number
      required: [totalDays, presentDays, absentDays, percentage]
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        targetRole:
          type: string
          enum: [all, admin, parents]
        createdAt:
          type: string
      required: [id, title, content, targetRole, createdAt]
...
    ChildProgress:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/Student"
        attendanceSummary:
          $ref: "#/components/schemas/AttendanceSummary"
        feeStatus:
          type: object
          properties:
            currentMonth:
              $ref: "#/components/schemas/FeeRecord"
            pendingCount:
              type: integer
            totalDue:
              type: integer
          required: [pendingCount, totalDue]
        recentAttendance:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
        announcements:
          type: array
          items:
            $ref: "#/components/schemas/Announcement"
      required: [student, attendanceSummary, feeStatus, recentAttendance, announcements]