# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Allied School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  rollNumber: string;
  name: string;
  class: string;
  /** @nullable */
  section?: string | null;
  fatherName: string;
  /** @nullable */
  motherName?: string | null;
  phone: string;
  /** @nullable */
  address?: string | null;
  admissionDate: string;
  monthlyFee: number;
  isActive: boolean;
  /** @nullable */
  leftDate?: string | null;
}