# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Allied School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  rollNumber: string;
  name: string;
  class: string;
  section?: string;
  fatherName: string;
  motherName?: string;
  phone: string;
  address?: string;
  admissionDate: string;
  monthlyFee: number;
}